+++ packages/mixte/index.ts
@@ -2,3 +2,4 @@ export * from './src/is/index';
 export * from './src/random/index';
 export * from './src/delay/index';
 export * from './src/uniqueKey/index';
+export * from './src/move/index';